+++ app/Http/Controllers/UserController.php
@@ -34,21 +34,24 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
-        //
-        $validated = $request->validated();
-
-         User::create([
-            'username' => $validated["username"],
-            'email'=> $validated['email'],
-            'password'=> bcrypt($validated['password']),
-            "role"=>$validated["role"]
-
-        ]);
-
-        
-        session()->flash('success', 'Formulir Pendaftaran Berhasil');
-        return redirect()->back();
+        try {
+            $validated = $request->validated();
+    
+            User::create([
+                'username' => $validated["username"],
+                'email' => $validated['email'],
+                'password' => bcrypt($validated['password']),
+                "role" => $validated["role"]
+            ]);
+    
+            session()->flash('success', 'Formulir Pendaftaran Berhasil');
+            return redirect()->back();
+        } catch (\Exception $e) {
+            session()->flash('error', 'Terjadi kesalahan saat menyimpan data.');
+            return redirect()->back();
+        }
     }
+    
 
     /**
      * Display the specified resource.
@@ -76,18 +79,23 @@ class UserController extends Controller
     public function update(UserRequest $request, string $id)
     {
         //
+        try{
 
-        $User = User::find($id);
-        $User->username = $request->username;
-        $User->email = $request->email;
-        $User->password =bcrypt($request->password) ;
-        $User->role = $request->role;
-
-        $User->save();
-
-        session()->flash('success', 'Formulir Profile User Berhasil di Update');
-
-        return redirect()->back();
+            $User = User::find($id);
+            $User->username = $request->username;
+            $User->email = $request->email;
+            $User->password =bcrypt($request->password) ;
+            $User->role = $request->role;
+    
+            $User->save();
+    
+            session()->flash('success', 'Formulir Profile User Berhasil di Update');
+    
+            return redirect()->back();
+        }catch(\Exception $e){
+            session()->flash('error', 'Terjadi kesalahan saat Update data.');
+            return redirect()->back();
+        }
     }
 
     /**
@@ -95,12 +103,22 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        //
-        $user = User::find($id);
-        $user->delete();
-
-        session()->flash('success', 'data User Profile User Berhasil di di delete');
-
-        return redirect('/user/create');
+        try {
+            $user = User::find($id);
+    
+            if (!$user) {
+                throw new \Exception('User tidak ditemukan.'); // Atau gunakan jenis Exception yang sesuai
+            }
+    
+            $user->delete();
+    
+            session()->flash('success', 'Data User Profile User Berhasil dihapus');
+            return redirect('/user/create');
+        } catch (\Exception $e) {
+            session()->flash('error', 'Terjadi kesalahan saat menghapus data: ' . $e->getMessage());
+    
+            return redirect('/user/create');
+        }
     }
+    
 }
